:snake: Make proxy/mock streams behave similarly
Now when streaming proxy responses the following happens in order;

0. Send status code
1. Iterate and send headers, if any
2. Close headers, if any
3. Iterate and send body lines, if any

This is different in that instead of attempting to re-send headers for
every line of response body, so _should_ be a little faster for Vim as
the channel callback will now not need to re-parse repeated headers.

Oh and this commit will be tracked by the next patch tagged version!